sendPOSI command change (double for lat/lon/h) (#111)
* Updated POSI to use doubles for lat/lon/alt, as step size for floats was unacceptably large at high longitudes.

+++ Java/Examples/BasicOperation/src/Main.java
@@ -25,11 +25,11 @@ public class Main
             xpc.getDREF("sim/test/test_float");
 
             System.out.println("Setting player aircraft position");
-            float[] posi = new float[] {37.524F, -122.06899F, 2500, 0, 0, 0, 1};
+            double[] posi = new double[] {37.524, -122.06899, 2500, 0, 0, 0, 1};
             xpc.sendPOSI(posi);
 
             System.out.println("Setting another aircraft position");
-            posi[0] = 37.52465F;
+            posi[0] = 37.52465;
             posi[4] = 20;
             xpc.sendPOSI(posi, 1);
